+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2026-06-19
+### Added
+- AzureRM provider features for `postgresql_flexible_server` and `virtual_machine`
+### Changed
+- `scripts/psql-init.sh` for full templating
+- `scripts/psql-init.sh` to run loop to valid Azure Key Vault access propagation status
+
 ## [0.1.0] - 2026-06-17
 ### Added
 - Initial commit